+++ src/shims/angular/angular.js
@@ -14,5 +14,12 @@
  * limitations under the License.
  */
 
-// Additional namespaces of the Kinvey service.
-var PUSH = 'push';
+// Angular.js.
+// -----------
+
+// Outside of the Kinvey module, use $q out of scope.
+var angularDeferred = angular.injector(['ng']).get('$q');
+Kinvey.Defer.use({ deferred: angularDeferred.defer });
+
+// Define the Angular.js Kinvey module.
+var module = angular.module('kinvey', []);